Roles functionality (#8)
According to the spec, these roles are required:  
- admin  
- team_member  
This could have been done with constants (or enums), and this approach should be considered if more roles are required. But as the spec only defines 2, a boolean flag is sufficient.  
Test checks that a new user will default to admin=false.